fix: move ResponseData to the Models folder
