+++ Modules/Devices/test/Devices.API.Tests.Integration/StepDefinitions/GenericResponseStepDefinitions.cs
@@ -1,3 +1,5 @@
+using Devices.API.Tests.Integration.Models;
+
 namespace Devices.API.Tests.Integration.StepDefinitions
 {
     [Binding]
